Add --local option for install script
To install clojure without downloading additional archive in case of some
specific distribution tooling (eg. yum and pacman)

+++ src/main/resources/install.sh
@@ -21,13 +21,14 @@ case "$(uname -s)" in
 esac
 
 default_prefix_dir="/usr/local"
+use_local=
 
-# use getopt if the number of params grows
 while [[ $# -gt 0 ]]
 do
     key="$1"
 
     case $key in
+        # path prefix where clojure will be installed
         -p|--prefix)
             if [[ -z "$2" ]]; then
                 do_usage
@@ -37,15 +38,26 @@ do
             shift # past argument
             shift # past value
             ;;
-        *)    # unknown option
+        # argument to skip downloading and expanding archive with the tool, install will use current dir instead.
+        # mainly for packaging needs so should not be mentioned in usage notes.
+        -l|--local)
+            use_local=true
+            shift # past argument
+            ;;
+        *) # unknown option
             do_usage
             ;;
     esac
 done
 
-echo "Downloading and expanding tar"
-curl -O https://download.clojure.org/install/clojure-tools-${project.version}.tar.gz
-tar xzf clojure-tools-${project.version}.tar.gz
+if [ $use_local = true ]; then
+    tools_dir="$(pwd)"
+else
+    tools_dir=clojure-tools
+    echo "Downloading and expanding tar"
+    curl -O https://download.clojure.org/install/clojure-tools-${project.version}.tar.gz
+    tar xzf clojure-tools-${project.version}.tar.gz
+fi
 
 lib_dir="$prefix_dir/lib"
 bin_dir="$prefix_dir/bin"
@@ -53,21 +65,23 @@ man_dir="$prefix_dir/share/man/man1"
 clojure_lib_dir="$lib_dir/clojure"
 
 echo "Installing libs into $clojure_lib_dir"
-$install -Dm644 clojure-tools/deps.edn "$clojure_lib_dir/deps.edn"
-$install -Dm644 clojure-tools/example-deps.edn "$clojure_lib_dir/example-deps.edn"
-$install -Dm644 clojure-tools/clojure-tools-${project.version}.jar "$clojure_lib_dir/libexec/clojure-tools-${project.version}.jar"
+$install -Dm644 "$tools_dir/deps.edn" "$clojure_lib_dir/deps.edn"
+$install -Dm644 "$tools_dir/example-deps.edn" "$clojure_lib_dir/example-deps.edn"
+$install -Dm644 "$tools_dir/clojure-tools-${project.version}.jar" "$clojure_lib_dir/libexec/clojure-tools-${project.version}.jar"
 
 echo "Installing clojure and clj into $bin_dir"
-$sed -i -e 's@PREFIX@'"$clojure_lib_dir"'@g' clojure-tools/clojure
-$install -Dm755 clojure-tools/clojure "$bin_dir/clojure"
-$install -Dm755 clojure-tools/clj "$bin_dir/clj"
+$sed -i -e 's@PREFIX@'"$clojure_lib_dir"'@g' "$tools_dir/clojure"
+$install -Dm755 "$tools_dir/clojure" "$bin_dir/clojure"
+$install -Dm755 "$tools_dir/clj" "$bin_dir/clj"
 
 echo "Installing man pages into $man_dir"
-$install -Dm644 clojure-tools/clojure.1 "$man_dir/clojure.1"
-$install -Dm644 clojure-tools/clj.1 "$man_dir/clj.1"
+$install -Dm644 "$tools_dir/clojure.1" "$man_dir/clojure.1"
+$install -Dm644 "$tools_dir/clj.1" "$man_dir/clj.1"
 
-echo "Removing download"
-rm -rf clojure-tools
-rm -rf clojure-tools-${project.version}.tar.gz
+if [ ! use_local = true ]; then
+    echo "Removing download"
+    rm -rf clojure-tools
+    rm -rf clojure-tools-${project.version}.tar.gz
+fi
 
 echo "Use clj -h for help."
